pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mattermost-mcp-host"
version = "0.1.0"
description = "Mattermost MCP Host with MCP Client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.13",
    "mattermost>=6.5.0",
    "mattermostdriver>=7.3.2",
    "mcp[cli]>=1.3.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.65.5",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
]

# Optional dependencies
[project.optional-dependencies]
anthropic = [
    "anthropic>=0.5.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
all = [
    "anthropic>=0.5.0",
    "google-generativeai>=0.3.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/mattermost-mcp-host"
"Bug Tracker" = "https://github.com/yourusername/mattermost-mcp-host/issues"

[project.scripts]
mattermost-mcp-host = "mattermost_mcp_host.main:main"